rename packages to align with configuration
