+++ dev/io.openliberty.data.internal/src/io/openliberty/data/internal/cdi/FutureEHFactory.java
@@ -51,8 +51,8 @@ import com.ibm.wsspi.persistence.DDLGenerationParticipant;
 import io.openliberty.data.internal.DataProvider;
 import io.openliberty.data.internal.EntityHandlerFactory;
 import io.openliberty.data.internal.Util;
-import io.openliberty.data.internal.provider.PUnitEHFactory;
-import io.openliberty.data.internal.service.DBStoreEHFactory;
+import io.openliberty.data.internal.ds.DBStoreEHFactory;
+import io.openliberty.data.internal.jpa.PUnitEHFactory;
 import jakarta.data.exceptions.DataException;
 import jakarta.persistence.EntityManagerFactory;
