+++ dev/io.openliberty.data.internal/src/io/openliberty/data/internal/jpa/package-info.java
@@ -15,6 +15,6 @@
  */
 @org.osgi.annotation.versioning.Version("1.0")
 @TraceOptions(traceGroup = "data", messageBundle = "io.openliberty.data.internal.resources.CWWKDMessages")
-package io.openliberty.data.internal.provider;
+package io.openliberty.data.internal.jpa;
 
 import com.ibm.websphere.ras.annotation.TraceOptions;
